Moved Task classes to separate files
