+++ src/main/java/signal/ToDo.java
@@ -1,0 +1,17 @@
+package signal;
+
+public class ToDo extends Task {
+    public ToDo(String description, boolean isDone) {
+        super(description, isDone);
+    }
+
+    @Override
+    public String toString() {
+        return "T" + super.toString();
+    }
+
+    @Override
+    public String checkType() {
+        return "ToDo";
+    }
+}
